+++ foxes/output/__init__.py
@@ -3,3 +3,4 @@ from .farm_layout import FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval
 from .flow_plots_2d import FlowPlots2D
 from .rose_plot import RosePlotOutput, AmbientRosePlotOutput
+from .results_writer import ResultsWriter