+++ foxes/output/farm_results_eval.py
@@ -2,7 +2,6 @@ import numpy as np
 import pandas as pd
 
 from .output import Output
-import foxes.constants as FC
 import foxes.variables as FV